+++ CHANGELOG.md
@@ -4,9 +4,9 @@
 
 ### Added
 
-- Added recommendations to the Blitz Cache utility.
+- Added recommendations to the Blitz cache utility.
 - Added new console commands for clearing, generating and purging the cache.
-- Added the source of actions and usernames to logging messages.
+- Added the user that runs controller actions to log messages.
 - Added an expiry date column to pages and elements in the Blitz Cache utility ([#615](https://github.com/putyourlightson/craft-blitz/issues/615)).
 
 ### Changed